requirements.txt
django==4.2.5
flake8==6.1.0
python-dotenv==1.0.0
pre-commit==3.5.0
flake8-bandit==4.1.1
bandit==1.7.5
psycopg2-binary==2.9.9
django-split-settings==1.2.0
django_debug_toolbar==4.2.0
pytest==7.4.3